Update patch set 7
Patch Set 7:

Yes, that's true. I'm on the fence because this is really the worst for tests that are bionic only. If we are planning to write a lot more bionic-only tests, then I'd lean towards generating once on a target. If we plan on writing tests to run on glibc too, and only the occasional bionic only test, then I'd say leave it this way since it will generate automatically.

Do you expect more bionic-only tests to come up, or more tests on glibc/bionic?

Patch-set: 7