Update patch set 5
Patch Set 5:

Updating the dynamic table is failing on N64 for a combination of reasons.

There is a difference in the linker scripts for O32 and N64 binaries. The N64 executables have the dynamic section in the first (read-only) loadable segment (this is where the dynamic header is normally placed for MIPS executables). In O32 the dynamic header was placed in the second (writable) load segment. I think this was an Android specific change.

binutils arbitrarily sets the flags of the DYNAMIC program header to RWX regardless of whether the segment is writable or not. There is a discussion going on in the binutils ML whether the flags have any meaning for a non-loadable program header.

I think the N64 linker script should be modified to put the dynamic table in the data segment in the same way as for N32. I'll work with Andrew to get this sorted out

Patch-set: 5
Reviewer: Gerrit User 1003284 <1003284@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0