Fix NSIG.
Since we switched to uapi headers, our sigset_t definition hasn't been tied
to our NSIG definition, so we can now fix it without breaking the LP32 ABI.
I don't think we can fix _NSIG though. That's both defined and used by the
uapi headers, so as long as we're exposing uapi headers to userspace we're
probably stuck with that.

One possibility: we could change the scripts to rename _NSIG to something like
_KERNEL_NSIG and then we could #define _NSIG to (_KERNEL_NSIG + 1) and then
both NSIG and _NSIG would have the appropriate userspace values.

Bug: 12938442
Change-Id: Ic7c86fd5be5ad1d822f7b2b1d88c8a0d70a1ac0f