Update patch set 3
Patch Set 3:

i saw the mksh reference but assumed we were taking a different branch (and only built arm). i'll just put sys_signame back. i was in two minds in the first place, since we're kind of a mix of glibc and BSD and people regularly try to build both kinds of source against us. if this were an Android-ism that would be different, but being in BSD is probably reason enough to support it, even if it's not in glibc.

Patch-set: 3