Fix race where SIGEV_THREAD timer locks can be taken after fork
While POSIX states that timers shouldn't carry over to the child,
bionic leaves the structures in place. This is useful, in case the
child keeps using the parent's timer_id:s (which is a POSIX violation).
In that case, if a SIGEV_THREAD timer lock is held during fork, some
timer operations (like timer_delete) will wait indefinitely for it.

This patch will ensure that child and parent will not share timer_id:s.
All timer calls will now fail if called with inherited timer_id:s,
implemented by an increasing "__timer_base_id" variable. A side effect
of this is that if you have 2^26 SIGEV_THREAD timer-creating ancestors,
you won't be able to create a SIGEV_THREAD timer. We've found, however,
that the practical limit on such ancestry is much lower than 2^26, with
or without this patch.

POSIX references:
The Open Group Base Specifications Issue 7, IEEE Std 1003.1, 2013 Ed.
 * fork(),         DESCRIPTION, section 1, bullet 17.
 * timer_create(), DESCRIPTION, section 4.

Co-authored-by: Snild Dolkow <snild.dolkow@sonymobile.com>

Change-Id: Ia2aa76147f4d4dc16134ca80644d743d77be70f2