Update patch set 1
Patch Set 1:

We had it happen in some internal code (not AOSP, not third-party). It was a C++ destructor trying to clean up. You know what C++ programmers are like when it comes to proper memory management... ;)

Since this is a race, it's easy to miss in testing, so I don't think the hanging will help anyone to locate any such bugs.

Besides, the timer's thread is gone, so the timer can't really be used anyway. It won't trigger any callbacks; anyone using inherited timer_ids for anything except timer_delete() should quickly notice that stuff just isn't working. With or without this patch.

The timer won't even free its thr_timer_table slot after it's deleted (because that would've been done in its thread). This also means that a child will be unable to create new SIGEV_THREAD timers if its ancestors used up all the slots, which this patch also fixes.

(This is my first time here. What's the etiquette on the "send email" check box?)

Patch-set: 1