Update patch set 1
Patch Set 1:

I'm not sure of the difference in the emitted instructions yet and I might be wrong but I don't think it makes a difference in terms of speed. Some claim that operations should be faster but I have no hard evidence of that. If any at all it will consume more memory bandwidth and thus might be a bit slower depending on what you factor in.

One case I can think of where this might be helpful is probably with alignment.

The other libc implementations I checked also use the register size for the fast ints.

Patch-set: 1