Allow SIGSEGV to be claimed by runtime
This change to the signal handling in bionic is to support our
use of the SIGSEGV signal in the android runtime to handle
null pointer checks and suspend checks.  The basic idea is that
the runtime can claim the SIGSEGV and any user calls to sigaction
that set the handler for SIGSEGV will be deferred.  This means
that the user's code will work as normal unless the SIGSEGV is generated
by our code.  If we don't generate the SIGSEGV, the user's handler
will be called from the signal handler and they won't know the
difference.

This also checks for the user attempting to block SIGSEGV using
sigprocmask.  We can't allow this to happen as it will prevent
the handling of null pointers and suspend checks by the runtime.

Bug: 13121132
Change-Id: I446c4745a28e42ba14886cafbd3439b6bb5eaada