Update patch set 2
Patch Set 2:

> I omitted the arpa_* files at first because they are referenced by
 > the upstream-netbsd/port_before.h which I thought it was an exact
 > replica.

no, the stuff in the top-level of each of those directories is our layer of porting hacks so we don't have to touch the source itself.

looking at this more closely, arpa_nameser.h should be arpa/nameser.h in the public headers. similar for nameser_compat.h. looks like we're a long way behind head with those files, though, so fixing that mess looks like it should be done in a separate change.

Patch-set: 2
Label: Verified=0