Update patch set 1
Patch Set 1:

It will likely be removed post-Volantis release.

It all stems from our ecosystem being cowardly comparatively. We have to support vendors that insist on deploying with the kernel logger, especially if they are retaining older kernels, do not want to disable the kernel logger for fear of change, and mixing them with later releases of the frameworks. The gTest I developed, I got pushback from test, security team and OEM-centric team members for making it an *error* that both loggers are deployed even though only one is being utilized. Apparently it is no ones responsibility to construct a CTS test that tells a vendor that they are being wasteful ...

I am braver than that, but I must listen to the voices from the team.

Patch-set: 1