Update patch set 1
Patch Set 1:

yeah, that was my comment on the original change: we should probably audit all this code. (or at least talk to upstream so they know they have a problem.)

can you either re-upload this with the same change-id as the original change or add a Signed-off-by so the original submitter gets some credit? otherwise, +2.

Patch-set: 1