Update patch set 1
Patch Set 1:

> For this and the other CLs: how do we decide which sources we pull
 > from which distributions?

that's a hard question. there are several competing criteria, including:

* which were we using previously (if upgrading)?

* which were we using previously for similar functions (if adding)?

* which has the best implementation?

* which fixes the bug/adds the feature we care about?

* which doesn't drag in as much random crap as the others?

* which is the "true" upstream? (FreeBSD seems to be the true upstream for libm, for example.)

if you look at the remaining BSD source that isn't currently in an upstream-* directory, you'll probably find that it raises an awkward question for at least one of these points. (the exceptions are ones where we actually need a patch relative to upstream, like stuff that touches struct fields that don't exist for Linux.)

Patch-set: 1