Update patch set 1
Patch Set 1: Verified+1

> Is there any way to indicate that the putw.c file is not strictly
 > following the upstream sources? I wouldn't want someone to
 > accidentally overwrite it with a newer version.

it is pure upstream, it's just a misleading diff. i replaced the FreeBSD one with the OpenBSD one. it's just a coincidence that i happened to also change the one file that the OpenBSD implementation #includes that the FreeBSD one doesn't. (why did i do it? because it gets rid of the "implicit declaration" warning we were getting from the FreeBSD source; the function in question is declared in "local.h".)

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1