Update patch set 2
Patch Set 2:

> (1 comment)

no, i think it's probably better as it is, for the reasons you mention.

the other alternative i thought of -- which i'm only bringing up in case we have future trouble building BSD code -- is to have more _USE_BSD in the headers so that BSD source can get more of the cruft while "glibc"/Linux source can get less.

Patch-set: 2