Update patch set 1
Patch Set 1:

(1 comment)

this seems unfortunate. we can't switch back to the default dispatch table around the call because it's a global and there might be multiple threads. couldn't we switch to a recursive mutex for the backlog?

or can we just build libgcc-demangle with references to dlrealloc et cetera, so it bypasses the checking, since it's used within the checking?

Patch-set: 1
Label: Verified=0