Update patch set 1
Patch Set 1:

(1 comment)

yeah, the fact that we have a weird/wrong implementation is why i didn't want to write tests. i'm assuming that we couldn't write tests that would pass on bionic and glibc (though on reflection that's probably not true as long as we only test ASCII values).

i'm hoping we'll find out soon what level of support for wchar_t we need for libc++. i'll have a look at writing tests for what we've got anyway, to see if the ASCII subset at least works like glibc.

Patch-set: 1
Label: Verified=0