Update patch set 9
Patch Set 9:

this passes the tests on arm and x86_64. i'll do some more testing in the morning, but i think this might be good now.

we should probably clean up the ieee.h stuff (and maybe have all architectures use the same one), but that can be done later.

Patch-set: 9
Label: Verified=0