Update patch set 1
Patch Set 1:

(1 comment)

short: I'm inclined to have this private if possible given that we have just one file.

I'm concerned that if we expose this and at some point we extend to an arch that's not compatible we'll have a hard time.  It probably won't happen any time soon but I feel that this "not gonna happen soon" lead to some bad stuff in the past.

Probably if a 3rd party project needs this is because we don't have some piece of functionality that they need. I guess we could just add that instead of exposing low level details.

Patch-set: 1