Update patch set 1
Patch Set 1:

yeah, it's kind-of implied in HACKING.txt but probably useful to explicitly spell it out there.

(i do wonder whether we'd be better off with a bionic-hacks/ directory [or similar] under each upstream-*, rather than using the top-level upsream-* directories for this directly. that might be self-documenting.)

Patch-set: 1