Handle special case of undefined symbol with defined st_value
This is how linker and dynamic loader resolve issue of getting address
of function defined in shared library for comparison.
ld.so on Linux has the same implementation.

Good description of the problem and fix is captured here:
http://www.airs.com/blog/archives/42

Note that issue cannot occur in PIE code. Test included into this change
is not very useful here because bionic tests are always built as PIE.
So it's included here just for reference.

Change-Id: I78e94e4a9e130849bd76c7fa8df92ad44aa9255f
Signed-off-by: Pavel Chupin <pavel.v.chupin@intel.com>