Switch to a working UTF-8 mb/wc implementation.
Although glibc gets by with an 8-byte mbstate_t, OpenBSD uses 12 bytes (of
the 128 bytes it reserves!). As long as we accept the limitation that the
caller needs to present us with a complete sequence before we'll process it,
we can actually implement UTF-8 encoding/decoding with a 0-byte mbstate_t
which means we can make things work on LP32 too. Another benefit is that
callers of the APIs that implicitly use a global mbstate_t all work fine
even in the presence of multiple threads.

Our behavior is fine when going from characters to bytes; we just
update the source wchar_t** to say how far through the input we got.

TODO: use the 4 bytes we do have to cope with byte sequences split across multiple input buffers?

Change-Id: Ied327a8c4643744b3611bf6bb005a9b389ba4c2f
TODO: add tests of valid and invalid UTF-8 sequences.