Update patch set 1
Patch Set 1:

OK, after reading the blog entry in its entirety (including comments), I do agree that the current bionic implementation violates the spec. Also, the described scenario indeed exists on ARM and Aarch64 as well.

However, I do suspect that this clashes with -Bsymbolic, as that results in the .so that defines the function not to refer to it through its symbol entry but through a pre-resolved relative relocation. So the only way to do this correctly is to disable -Bsymbolic for those non-PIE builds.

Patch-set: 1