Update patch set 1
Patch Set 1:

I don't have a test case prepared, but I can explain why your change -while
correct at first glance- is a bad idea for Bionic.

The Android ecosystem relies heavily on JNI shared libraries that are hosted by
Dalvik VM instances spawned from zygote. Those libraries are usually built with
-Bsymbolic, but as there is no 'legacy' executable in play, the issue your patch
solves presents itself rarely.

Now consider what will happen when we apply your patch: all of a sudden,
libraries that take the addresses of each others' functions will get different
results depending on whether they are needed by an executable that takes the
address of any of these functions as well.

For instance, liba.so contains foo() and libb.so takes its address. In the old
situation, both liba and libb will get the same address: liba because it uses a
relative relocation generated at link time, and libb because resolving its
GLOB_DAT relocation of symbol 'foo' produces the real address of 'foo'.

After applying your patch, the above will still apply *unless* the executable
that loads liba and libb takes the address of 'foo' as well. In that case, the
reference libb has will be preempted, but liba's internal reference will still
point to the real place, resulting in different values from the pov of liba and
libb. Note that liba.so and libb.so will be the exact same binaries, the only
difference is in the executable that needs them.

My personal opinion is that this patch is not necessary, but if others feel that
it is, you should at least update it so the linker errors out on encountering
such a relocation against a symbol that is defined in a library built with
-Bsymbolic.

Patch-set: 1
Reviewer: Gerrit User 1014784 <1014784@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0