Update patch set 1
Patch Set 1:

I still think that what you are describing is irrelevant to this patch. It changes anything for JUMP_SLOT only. Case you are describing is very easy to build and try. Probably I misunderstand something but I tried to build it before and after patch and nothing changes.

Anyway, if we really don't care about non-PIE executable anymore this patch is probably redundant.

Patch-set: 1