Update patch set 2
Patch Set 2:

(1 comment)

Constructor used to rely on bname to find already loaded libraries and this is not applicable in the case where same library (based on stat) has different names. namely symlinks.

This is why we need dependencies to be represented in soinfo prior to CallConstructors. Also they will help with fixing dlsym but this is subject to separate change.

Here is example use case:
1. binary DT_NEEDs libexample.1.so
2. binary DT_NEEDs libother.so -> DT_NEEDS libexample.so -> symlink to libexample.1.so

resolving by name will lead to an error. because soinfo.name for both example.1 and example are libexample.1.so and it cannot be resolved using existing find_loaded_library with libexample.so. It will need to repeat file-system dependent code, which may fail. Not that CallConstructors cannot handle any errors.

Patch-set: 2