Update patch set 6
Patch Set 6:

ptal.

I did not get to the bottom of glibc behavior but I clean up the tests and I think we're good so far. It is possible to get a fpos_t which is inside a muti byte (also in glibc) but then, as expected, reading the next multi byte will fail.

I still suspect a bug in glibc/libio though looking at the code did not reveal anything suspicious yet...
- if I close the file and reopen it in read more then everything works as expected. 
- if I don't use setlocale(...,"C.UTF-8") the output is almost as expected except that fgetwc doesn't set the errno anymore (the current locale on my testing system is en_US.UTF-8)

Patch-set: 6