Update patch set 1
Patch Set 1:

There's still missing stuff on arm but I'm not sure what's the best source to get them. Especially for aarch64 which doesn't have any space for the non core registers. Glibc seems to have incomplete structs as well and mentions that non core regs should be stored on the stack or in the uc_regspace (which I assume is just an unstructured reservation?)

Patch-set: 1