Update patch set 1
Patch Set 1:

Looking from the sync-up perspective: The resolver already divergence in a consistent way from the upstream so it might be better to take ownership. I found quite difficult to figure out what should be synced  when I was looking into this (note that the sync is not fully done, see below). It would be nice to be consistent with the upstream but it's getting harder to sync when it's not clear what we use and what we don't. So increasing the androidness by cleaning stuff might actually make things easier (wrt code understanding and taking upstream patches).

Either way, our resolv sources are pretty outdated compared with the upstream. It would be best to finish the sync before anything. We had a partial sync up tracked here b/13219633. I stopped to wait for https://android-review.googlesource.com/#/c/82360 but then I didn't find the time to come back and finish it. What's missing is basically the resolv stuff.

Patch-set: 1
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0