Update patch set 1
Patch Set 1:

the trouble is that this is private API. and, as you said on the previous review, none of the public API callers even look at the non-failure return values.

one possibility is that maybe we should just change the private API to be less dangerous. the only reason i'm not looking at switching everyone over to something that sets errno and has a regular syscall-like return is that almost all the callers are in the pthread implementation, so they'd all want to save/restore errno.

Patch-set: 1