Update patch set 1
Patch Set 1:

bzero can actually be a couple of cycles faster than memset 0 (e.g. in arch-arm64/generic/bionic/memset.S saves a call to ands A_lw, val, #255 because the value is guaranteed to be 0), so if we want to save every possible CPU power, bzero can stay (OTOH its presence makes code bigger).

bcopy is entirely useless though, it's never faster than memmove and should be #define-d away for performance.

I'm fairly certain glibc puts it in as a symbol only because standards say it's a function.

Patch-set: 1