Update patch set 3
Patch Set 3:

this is the kind of change where you'll have to go round and fix the callers first. it looks like a lot of the references are for Windows builds (and not our problem). you'll have to checkbuild to find which references we do need to fix.

Patch-set: 3
Label: Verified=0