Update patch set 4
Patch Set 4:

I also think that it's not great to leave vfork as vfork on LP32 and have vfork as fork on LP64 (even if we don't have a symbol for it). That's for the same reasons Elliot mentioned. Fair enough that it should be the programmer's responsibility to check and make sure the code is compatible but we know that doesn't usually happen...

However this is also what the BSD has done according to the man pages ("In 4.4BSD it was made synonymous to fork(2) but NetBSD introduced it again, cf. http://www.netbsd.org/Documentation/kernel/vfork.html").

Maybe we should be bold and just remove any declaration from LP64 and keep it as the real vfork on LP32 if POSIX <= 2008.

(fyi, glibc 2.15 exposes vfork as a weak symbol that has not been specifically tagged as a weak object symbol)

Patch-set: 4
Label: Code-Review=0