Update patch set 4
Patch Set 4:

I like the be Cailin's bold idea. Specifically, it seems like it might be a good idea to:

On LP32, keep the vfork prototype if POSIX < 200809. Keep the vfork implementation as is.

Remove it completely from LP64 and LP32 where POSIX >= 200809.

But do we have to support POSIX < 200809, or can we say that it's not supported at all? That way we could move the vfork to ndk_cruft for LP32 and remove any mention of vfork from the rest of the system. I'm not sure what line can be drawn about support like this though.

Patch-set: 4
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0