Get the full set of PRNG functions in <stdlib.h>
Note that NetBSD gets some of the types wrong. random.c here is slightly
modified, so we should pull it out of upstream before submitting. I've
reported this upstream to see if they're likely to fix it.

FreeBSD has the same problem. OpenBSD is correct, but they have some
BSD-specific code in their implementation to get a good random default
seed from the kernel.

Change-Id: I435250bdae302e8bd7e29977d0fde7b9afbfca5e