libc: Fix 'index' symbol export.
The C library didn't export the 'index' symbol, but its C++ name-mangling
instead, which broke the ABI and prevented some applications from loading
properly.

The main reason was that the implementation under bionic/index.cpp relied
on the declaration to specify that the function has C linkage.

However, the declaration for index() was removed from both <string.h>
and <strings.h> in a recent patch, which made the compiler think it was
ok to compile the function with C++ linkage instead!

This patch does the following:

- Add back index() in <strings.h>, it's POSIX-compliant
  location, with a note explaining why it can't be declared as a macro.

  This is to avoid breaking the compilation of third-party applications
  after the next NDK header refresh.

- Add back index() in <string.h>, to prevent breaking the compilation of
  third-party code that mistakenly includes this header, instead of
  <strings.h> to get it.

- Convert the implementation source into a C file by changing its
  extension to avoid any similar surprises in the future.

BUG=15606653

Change-Id: I816b589f69c8f8a6511f6be6195d20cf1c4e8123