Update patch set 3
Patch Set 3:

Another option I discussed with Paul was to create a board configuration option, and only enable this support for hardware which explicitly opts into it. We could then enable this for Nakasi, but continue to leave it removed for other hardware.

But the best case is that we get an updated binary, so we can remove this code once and for all.

Patch-set: 3