Cache getpid.
In practice, with this implementation we actually only need to make at
most one system call, the first time getpid is called. A new thread will
have the same pid as its parent, and a fork child's main thread will have
a pid equal to its tid, which we get for free from the kernel before
clone returns.

Change-Id: I6d4718ed0a5c912fc75b5f738c49a023dbed5189