Update patch set 1
Patch Set 1:

> It seems to me that the likely core issue here is that if a key is
 > deleted and reallocated, we need to clear corresponding per-thread
 > data for all threads, so that values for the new key appear to be
 > NULL?  That seems tricky.  I didn't check whether the old code is
 > correct.
 
correct. the old code has a global list of all the threads and runs down that list. the old code just assumes that any read that happens after the write gets the new value.

(i suspect that this is mostly theoretical in practice. it's highly unlikely any code is (a) recycling keys and (b) taking advantage of the fact that they're supposed to be initialized to NULL.)

 > I think there is also a race if a thread uses a newly created key
 > and exits before the destructor is installed.
 > 
 > The performance issues seem to have been anticipated by the Posix
 > spec, in that they discourage checking key validity, esp. in
 > pthread_getspecific().  The old validity check did not enforce
 > memory ordering, and was thus not too expensive, though probably
 > still undesirable.  The new one enforces memory ordering, making it
 > much worse.  That could be fixed by either dropping the expensive
 > part of the check or making it memory_order_relaxed.
 > 
 > I'm not sure I would go in this direction at all.  But I don't
 > think the observed performance problem is the main reason for that.
 >  If there is a strong reason to do this, I would merge the
 > destructor and in_use bit, using a special value for (no destructor
 > but in use), which may also be useful while clearing data across
 > all threads.

the fewer global data structures we have, the better. but in this case i'm struggling to think of a way to make the common call (pthread_getspecific) cheap and pushing the expenses into the rare calls (pthread_key_create, pthread_key_delete, pthread_setspecific).

the only other thing i can think of is to actually have a sequence number. but that doubles the space needed for the TLS in every thread (LP32; 4/3 in LP64), and i'm not sure i'm comfortable with that.

Patch-set: 1