Under LP64 configuration, full name instead of based is used after commit 489e498434f53269c44e3c13039eb630e86e1fd9.
But when checking whether the depended DSO is loaded or not and
if the depended DSO is not in default search dir, loader should
just check the base name of soinfo.name. Because there is no
full name recoarded for depended DSO.

Example. Let's say we have two libs: liba.so and libb.so
liba.so is NEEDED by libb.so. And they all are in dir:
/system/lib64/xxxx/

User use dlopen to load these two libs like:
  dlopen(/system/lib64/xxxx/liba.so);
  ...
  dlopen(/system/lib64/xxxx/libb.so);
    ---> fails here with error like liba.so was not loaded.

Side-effect: if several DSO has the same base name, that could
confuse the symbol lookup.

Change-Id: I4fff3e5961da65b048b2232f631b37fbb8869bab
Signed-off-by: Qiming Shi <qiming.shi@intel.com>
Signed-off-by: Yin, Fengwei <fengwei.yin@intel.com>