Update patch set 1
Patch Set 1:

I'm a bit confused about the generic/generic-neon switch - I thought neon was mandatory in v8, but apparently there's a need for a non-neon branch anyway? Is it just about CPUs that are slow at executing neon?

I was assuming this should be generic-neon primarily because of the use of cmeq and the comment in the original code
/* Assumptions:
 *
 * ARMv8-a, AArch64
 * Neon Available.
 */

Will upload a new revision that moves strchr.S to generic, but it would be good to know what the non-neon split is about for future reference.

Patch-set: 1
Reviewer: Gerrit User 1010121 <1010121@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0