Update patch set 2
Patch Set 2:

Feel free to ignore generic-neon for now. I created it as a place to hold certain version of the functions that use neon instructions in the main loop. I should probably have created a better name for it, or just put it in generic. I'll clean this up at some point in the near future.

Patch-set: 2