Do not include libc_common in malloc debug code.
The inclusion of the static libc_common library in the malloc_debug_XXX.so
shared libraries causes a double init. This doesn't seem to have caused
any issues when setting the libc.debug.malloc property. However, jemalloc
crashes sometimes when this mix occurs. It's possible this causes problems
but that nobody has found them yet.

Change-Id: Iaac650a82d69064db148a6333e9403744f68b4a4