Update patch set 1
Patch Set 1:

there were two problems on x86-64.

one is that there are #pragmas in the x86/x86-64 code that neither gcc nor clang recognize, so we need to put the -Wno-unused-pragmas back. that's easily fixed and i'll upload ps2 as soon as i finish typing this.

the second is that the platform clang seems to be configured for ld80 long doubles rather than ld128? (the intention for LP64 is to have all the platforms use ld128 long doubles, and that's what the NDK currently does.)

Patch-set: 1
Label: Verified=0