Update patch set 1
Patch Set 1:

> Hi Elliott,
 > 
 > Wee see that gcc.dg/c99-stdint-1.c,  gcc.dg/c99-stdint-7.c tests
 > fail on Linux. Those tests, among other things, check:
 > 
 > 7.18.3
 > If wint_t (see 7.24) is defned as a signed integer type, he value
 > of WINT_MIN shall be no greater than -32767 and the value of
 > WINT_MAX shall be no less than 32767; otherwise,  wint_t is defned
 > as an unsigned integer type, and the value of WINT_MIN shall be 0
 > and the value of WINT_MAX shall be no less than 65535.
 > 
 > 
 > However, when __WINT_UNSIGNED__ is defined, those test pass. Where
 > this macro has to be defined? I see that clang defines it, should
 > gcc also define it as well?

it's not obvious to me what part of 7.18.3 this patch doesn't follow? are you saying that you have an unsigned wint_t but don't define __WINT_UNSIGNED__?

(afaics, glibc unconditionally assumes that wint_t is unsigned and the BSDs unconditionally assume that wint_t is signed. iirc we try to cope with both because x86 differs from the other two?)

Patch-set: 1