Update patch set 1
Patch Set 1:

It's an ABI difference. The SLM has the instruction so doesn't need the cruft. The NDK ABI... however... doesn't include the POPCNT instruction and the compiler calls into a helper routine.

if libc.so and libgcc.a are built for slm... we need to ensure that we continue to include the helper routines (even if the library itself doesn't need them).

Patch-set: 1