Added test for ifunc support in dynamic linker.
ifuncs now work in x86_64 when called in the same library as well as in a
different library.

In i386, they don't work yet because it appears that there is a
R_386_IRELATIVE reloc type that I need to deal with.  But I wanted to
upload the working 64 bit for now.

Bug:6657325
Change-Id: Ic0c48b1b0a76cb90f36c20c79f68294cc3fd44a1