[WIP] Added test for ifunc support in dynamic linker.
STT_GNU_IFUNC symbols now work properly.  Instead of returning the
pointer to the selection function, the linker instead now runs that
function and then returns the address of the function chosen.

In addition, LD_DEBUG output now detects and prints symbol type on a
lookup operation and exits when an unsupported symbol type is detected.

Bug:6657325
Change-Id: Ic0c48b1b0a76cb90f36c20c79f68294cc3fd44a1