[WIP] Added test for ifunc support in dynamic linker.
This doesn't work properly, and it is either the consequence of me
misusing pointers (perfectly possible) or the linker not quite doing its
job properly.  The tests fail in this case because foo, which is
properly getting redirected to foo_ifunc, is
returning the address of the function to be called (i.e. f1 or f2)
instead of calling that function and returning its value.  If line 34 of
dlopen_testlib_ifunc.c is changed to returning f1() or f2() instead of
the function pointer, everything works as expected (but not like an
ifunc should).

Before I spend any more time on this trying to figure out what's wrong,
I wanted to submit a CL and see if I'm just making some simple error
here.  I don't see any sort of R_386_IRELATIVE relocation within the .so
either, so I am inclined to believe something is amiss beyond what I'm
doing.

Bug:6657325
Change-Id: Ic0c48b1b0a76cb90f36c20c79f68294cc3fd44a1