[WIP] Added test for ifunc support in dynamic linker.
THIS DOESN'T WORK RIGHT NOW!  I get a SEGV_ACCERR on line 525 of
linker.cpp, which is where I attempt to update the ElfW(Sym) st_value.
This is the same error I was getting before, though now I believe I've
moved the actual invocation of this function to the right location.  Can
I change the value of st_value ever, or do I need to do something like
replace it with a copy?

Bug:6657325
Change-Id: Ic0c48b1b0a76cb90f36c20c79f68294cc3fd44a1

Conflicts:
	tests/libs/Android.mk