Added test for ifunc support in dynamic linker.
Right now this half-works.  Getting the actual ifunc from dlsym works,
but running the ifunc within the library does not (it just returns the
pointer to the function that should be called).  So I think I'm on the
right track and need to figure out what happens differently between the
two cases.

Bug:6657325
Change-Id: Ic0c48b1b0a76cb90f36c20c79f68294cc3fd44a1