Update patch set 3
Patch Set 3:

please take a look at patchset #3. This change adds the logging requested in patchset 2. In addition, this change modifies the FORTIFY_SOURCE tests to suppress debuggerd logging on the expected crash of the application. This reduces the total runtime of the bionic unittests by over 50%

Patch-set: 3
Label: Verified=0