Use libunwindbacktrace for debug malloc code.
Add a file containing replacements for functions that do allocations either
directly or indirectly so that the libunwind code can be pulled in without
any changes.

Also, remove the debug_XXX code that doesn't need to be in libc.so.

Fix the maps reading code since it was completely broken for 64 bit.

Bug: 16408686

Change-Id: I6b02ef6ce26fdb7a59ad1029e7cbba9accceb704