Update patch set 6
Patch Set 6:

Crystal clear, it is a good reason. I agree to abandon this patch and using Varvara's solution. 
it is a bad way using hard code in script, it is because we haven't found the better solution to automatically detect the missed symbol as a result of popcnt instruction support. I thought directly added the compat file in silvermont folder and modify the .mk file. instead, I choose the current solution, in order to align with arm, beside, I thought not merely for silvermont arch, but for future archs support popcnt instruction.

Patch-set: 6
Label: Verified=0