Update patch set 2
Patch Set 2:

The shims are so the libunwind code can call free/malloc/calloc/strdup and avoid an infinite loop (or really getting stuck in a mutex). These functions get linked with the statically linked libunwindbacktrace.a and resolve those functions so that libunwind doesn't call any memory allocation functions.

I needed to add the FILE* functions because they can do an allocation too and are used in libunwind and the debug_mapinfo.cpp code.

Patch-set: 2