Update patch set 10
Patch Set 10:

If it is possible to just update not-shipped library - lets do that. 

This really should not be a problem for arch's other than arm, afaik link order was fixed on more recent versions of ndk - so we shouldn't need compat for x86.

Patch-set: 10
Reviewer: Gerrit User 1041673 <1041673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0