fts: optimize by using fstatat instead of lstat
This is part of an effort to improve time it takes
to restorecon on /sys.

During file traversal use existing file descriptor
when doing stat. This improves fts_read with 40-50%.

Test case measured on Z Ultra GPE:

strace -c restorecon -r /sys

Original:

 % time     seconds  usecs/call     calls    errors syscall
 ------ ----------- ----------- --------- --------- ----------------
 47.24    0.001283           1      1065           newfstatat
 21.72    0.000590           2       265           lgetxattr
 18.52    0.000503          84         6           prctl

With the fix:

 % time     seconds  usecs/call     calls    errors syscall
 ------ ----------- ----------- --------- --------- ----------------
 47.36    0.000708           1      1065           newfstatat
 40.13    0.000600           2       265           lgetxattr
 12.51    0.000187           4        48           mprotect

glibc is using the same optimization.

Change-Id: Ia7e26b603f6461095b5a8d7e8a6cdf4f8279fa84