Create a distinct temp directory for each run.
Also, modify make__NR_name so that starting a name with __ does not
cause a name to be used without __NR_ being prepended.  Instead, a
check for __ and finding _NR_ in the name results in no prepend. This
avoids a case where using a name starting with __ in SYSCALLS.TXT
does not generate code that will compile but references the function
itself and causes link errors.

Change-Id: Ib9527eba6f25f26a30c5cb0ad431f3f88a7683cf